# swagger.yaml
openapi: 3.0.3
info:
  title: Kakarotos Webpage
  description: Esta API permite acceder y añadir información relacionada con el equipo Kakarotos FC. Permite agregar información personal de los jugadores e información de rendimiento a nivel individual y colectivo. También permite acceder a ella a traves de búsquedas individuales o del equipo. Por último, cuenta con una tienda donde se encuentran los productos asociados con la marca detrás del equipo (uniformes completos o piezas individuales).
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
  /player/save:
    post:
      tags:
        - player
      summary: Inserta o actualiza la información de un nuevo registro en la entidad Player. 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Player'
      responses:
        200:
          description: Entidad guardada exitosamente.
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/ServerError'
...
  /player/list_active_players:
    get:
      tags:
        - player
      summary: Obtiene y devuelve una lista de todos los registros de la entidad Player que cumplan con la condición departure_date = null. Si el registro tiene algún valor distinto, el endpoint no lo devuelve.
      responses:
        200: 
          description: (Ok) El endpoint devuelve al menos un registro exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
...
components:
  responses:
    BadRequest:
      description: (Bad Request) no es un formato de mensaje que cumpla con el tipo de dato para alguno de los parámetros. 
    NotFound:
      description: (Not Found) no se encuentra ningún registro en la entidad que se consulta o los parámetros enviados no corresponden a ningún registro en específico.
    ServerError:
      description: Error en el servidor.
  schemas:
    Player:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        number:
          type: integer
        position:
          type: string
        foot:
          type: string
          maxLength: 1
        weakFoot:
          type: integer
        skills:
          type: integer
        startDate:
          type: string
          format: date
        departureDate:
          type: string
          format: date
        imageUrl:
          type: string
...